+++ pg/networks.py
@@ -1,0 +1,26 @@
+import torch
+import torch.nn as nn
+from torch.distributions import Categorical
+
+class PolicyNetwork(nn.Module):
+    def __init__(self, input_dim, hidden_dim, output_dim):
+        super(PolicyNetwork, self).__init__()
+        self.fc1 = nn.Linear(input_dim, hidden_dim)
+        self.fc2 = nn.Linear(hidden_dim, output_dim)
+
+    def forward(self, x):
+        x = torch.relu(self.fc1(x))
+        x = self.fc2(x)
+        return Categorical(logits=x) # when inferencing no need to call Categorical again
+
+# If we think of the policy as an actor, value function is the "critic"
+# This, thus, makes a simple actor-critic architecture
+class ValueNetwork(nn.Module):
+    def __init__(self, input_dim, hidden_dim):
+        super(ValueNetwork, self).__init__()
+        self.fc1 = nn.Linear(input_dim, hidden_dim)
+        self.fc2 = nn.Linear(hidden_dim, 1)
+
+    def forward(self, x):
+        x = torch.relu(self.fc1(x))
+        return self.fc2(x)
