+++ pg/notes/vanilla-pg.md
